Fix the tenth FMA site (compound assignment); drop the lint per ruling
Codex round 3 (P1, EXECUTED): the lint only inspected BinaryExpr
ancestors, never AssignStmt, so compound assignment (+=/-=) bypassed
it entirely. testops_risk_native_clickhouse.go's computePipelineStability
had three live unguarded sites this way -- successRate7d, num, den --
found by go tool objdump (FMADDD/FMSUBD present) after the reported one
was fixed; all three wrapped in float64() and verified removed from the
compiled arm64 binary. This is a DIFFERENT defect from the CHAOS-4824
compensated-sum() issue already deferred to PR #2107 on this same
accumulator: a per-step FMA fusion present regardless of summation
strategy.

Bit-pattern test added (STORED/rounded value, since this branch does not
have PR #2107's unrounded-value extraction) -- disclosed honestly in the
PR body that it does not reliably distinguish fused/unfused for this
site's small corpus (divergences this class produces rarely survive
4-decimal rounding, matching the general finding already documented for
the sibling CHAOS-4824 PR). The primary, executed proof for this fix is
the disassembly: FMADDD/FMSUBD present before, absent after.

Per team-lead's ruling: dropped the AST lint (internal/jobs/fma_lint_test.go)
from this PR entirely -- it has a real class of blind spots beyond
compound assignment (any statement form that can hold a product feeding
an add elsewhere) and ships as its own follow-up PR once it covers them,
rather than blocking the nine (now ten) site fixes and their bit-pattern
tests, which are independently proven and unaffected by the lint's gaps.

+++ internal/jobs/metrics/daily/testops_risk_native_clickhouse.go
@@ -827,7 +827,16 @@ func computePipelineStability(repoID uuid.UUID, day time.Time, dayEntries []test
 	}
 	successRate7d := 0.0
 	for i, m := range dayEntries {
-		successRate7d += m.SuccessRate * weights[i]
+		// float64(...) is load-bearing (CHAOS-4818, codex round 3 on
+		// PR #2106): compound assignment desugars to
+		// `successRate7d = successRate7d + m.SuccessRate*weights[i]`, and Go
+		// may fuse the product into that addition on arm64 -- a DIFFERENT
+		// defect from the CHAOS-4824 compensated-sum() issue on this same
+		// accumulator (tracked separately, PR #2107): this is a per-step FMA
+		// fusion present regardless of whether the accumulation strategy is
+		// naive or compensated. Measured: `go tool objdump` showed FMADDD at
+		// this line before the wrap.
+		successRate7d += float64(m.SuccessRate * weights[i])
 	}
 	successRate7d /= totalWeight
 
@@ -842,8 +851,13 @@ func computePipelineStability(repoID uuid.UUID, day time.Time, dayEntries []test
 		num := 0.0
 		den := 0.0
 		for i, m := range dayEntries {
-			num += (float64(i) - xMean) * (m.SuccessRate - yMean)
-			den += (float64(i) - xMean) * (float64(i) - xMean)
+			// float64(...) is load-bearing (CHAOS-4818, codex round 3 on
+			// PR #2106): found by disassembly while verifying the :830 fix
+			// -- these two compound assignments have the SAME live FMADDD/
+			// FMSUBD exposure as successRate7d above, for the same reason
+			// (compound assignment desugars to a fusable add/subtract).
+			num += float64((float64(i) - xMean) * (m.SuccessRate - yMean))
+			den += float64((float64(i) - xMean) * (float64(i) - xMean))
 		}
 		if den > 0 {
 			successRateTrend = num / den
